Refactor/model classes (#41)
* Deleted unused model classes

* Moved ChangeEmailRequests model class out of model.py

* Finished to move ChangeEmailRequests model class out of model.py

* Moved GlobalData model class out of model.py

* Moved LoginAttempts model class out of model.py

* Moved PasswordReset model class out of model.py

* Moved RevokedTokenModel model class out of model.py

* Moved SendMailAttempts model class out of model.py

* Moved StatsCounters model class out of model.py

* Restored wrongly deleted model class

* Fixed the way of importing the moved model classes
