+++ iot_api/user_api/model.py
@@ -273,22 +273,6 @@ class UserToUserRole(db.Model):
         return cls.query.filter_by(user_id=user_id, user_role_id=user_role_id).first()
 
 
-class RevokedTokenModel(db.Model):
-    __tablename__ = "revoked_tokens"
-
-    id = db.Column(db.BigInteger, primary_key=True)
-    jti = db.Column(db.String(120))
-
-    def add(self):
-        db.session.add(self)
-        db.session.commit()
-
-    @classmethod
-    def is_jti_blacklisted(cls, jti):
-        query = cls.query.filter_by(jti=jti).first()
-        return bool(query)
-
-
 class AccountActivation(db.Model):
     __tablename__ = "account_activation"
 
@@ -334,120 +318,6 @@ class AccountActivation(db.Model):
         return cls.query.filter_by(user_id=user_id).order_by(desc(AccountActivation.creation_date))
 
 
-class PasswordReset(db.Model):
-    __tablename__ = "password_reset"
-
-    id = db.Column(db.BigInteger, primary_key=True)
-    user_id = db.Column(db.BigInteger, db.ForeignKey(
-        "iot_user.id"), nullable=False)
-    token = db.Column(db.String(500), nullable=False)
-    creation_date = db.Column(db.DateTime(timezone=True), nullable=False)
-    active = db.Column(db.Boolean, nullable=False, default=True)
-
-    def to_json(self):
-        return {
-            'id': self.id,
-            'user_id': self.user_id,
-            'token': self.token,
-            'creation_date': self.creation_date,
-            'active': self.active
-        }
-
-    def save_to_db(self):
-        db.session.add(self)
-        db.session.commit()
-
-    def update_to_db(self):
-        db.session.commit()
-
-    def delete_from_db(self):
-        db.session.delete(self)
-        db.session.commit()
-
-    @classmethod
-    def find_by_token(cls, token):
-        return cls.query.filter_by(token=token, active=True).first()
-
-    @classmethod
-    def find_active_tokens_by_user_id(cls, user_id):
-        return cls.query.filter_by(user_id=user_id, active=True).all()
-    
-
-class LoginAttempts(db.Model):
-    __tablename__ = "login_attempts"
-
-    id = db.Column(db.BigInteger, primary_key=True)
-    user_id = db.Column(db.BigInteger, db.ForeignKey("iot_user.id"), nullable=False)
-    attempts = db.Column(db.Integer, nullable=False, default=True)
-    last_attempt = db.Column(db.DateTime(timezone=True), nullable=False)
-
-    def to_json(self):
-        return {
-            'id': self.id,
-            'user_id': self.user_id,
-            'attempts': self.attempts,
-            'last_attempt': self.last_attempt
-        }
-
-    def save_to_db(self):
-        db.session.add(self)
-        db.session.commit()
-
-    def update_to_db(self):
-        db.session.commit()
-
-    def delete_from_db(self):
-        db.session.delete(self)
-        db.session.commit()
-
-    @classmethod
-    def find_by_user(cls, user_id):
-        return cls.query.filter_by(user_id=user_id).first()
-
-
-class ChangeEmailRequests(db.Model):
-    __tablename__ = "change_email_requests"
-
-    id = db.Column(db.BigInteger, primary_key=True)
-    user_id = db.Column(db.BigInteger, db.ForeignKey(
-        "iot_user.id"), nullable=False)
-    new_email = db.Column(db.String(120), nullable=False)
-    old_email = db.Column(db.String(120), nullable=False)
-    token = db.Column(db.String(500), nullable=False)
-    creation_date = db.Column(db.DateTime(timezone=True), nullable=False)
-    active = db.Column(db.Boolean, nullable=False, default=True)
-
-    def to_json(self):
-        return {
-            'id': self.id,
-            'user_id': self.user_id,
-            'new_email': self.new_email,
-            'old_email': self.new_email,
-            'token': self.token,
-            'creation_date': self.creation_date,
-            'active': self.active
-        }
-
-    def save_to_db(self):
-        db.session.add(self)
-        db.session.commit()
-
-    def update_to_db(self):
-        db.session.commit()
-
-    def delete_from_db(self):
-        db.session.delete(self)
-        db.session.commit()
-
-    @classmethod
-    def find_by_token(cls, token):
-        return cls.query.filter_by(token=token, active=True).first()
-
-    @classmethod
-    def find_active_tokens_by_user_id(cls, user_id):
-        return cls.query.filter_by(user_id=user_id, active=True).all()
-
-        
 class AlertType(db.Model):
     __tablename__ = 'alert_type'
     id = Column(BigInteger, primary_key=True, autoincrement=True)
@@ -967,23 +837,10 @@ class Device(db.Model):
         return query.all()
     
 
-class DevNonce(db.Model):
-    __tablename__ = 'dev_nonce'
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    dev_nonce = Column(Integer, nullable=True)
-    device_id = Column(BigInteger, db.ForeignKey("device.id"), nullable=False)
-    packet_id = Column(BigInteger, db.ForeignKey("packet.id"), nullable=False)
-
-
 class GatewayToDevice(db.Model):
     __tablename__ = 'gateway_to_device'
     gateway_id = Column(BigInteger, db.ForeignKey("gateway.id"), nullable=False, primary_key=True)
     device_id = Column(BigInteger, db.ForeignKey("device.id"), nullable=False, primary_key=True)
-
-class GatewayToDeviceSession(db.Model):
-    __tablename__ = 'gateway_to_device_session'
-    gateway_id = Column(BigInteger, db.ForeignKey("gateway.id"), nullable=False, primary_key=True)
-    device_session_id = Column(BigInteger, db.ForeignKey("device_session.id"), nullable=False, primary_key=True)
 
 
 class DeviceSession(db.Model):
@@ -1158,88 +1015,6 @@ class DeviceAuthData(db.Model):
     device_session_id = Column(BigInteger, ForeignKey("device_session.id"), nullable=True)
 
 
-class RowProcessed(db.Model):
-    __tablename__ = 'row_processed'
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    last_row = Column(Integer, nullable=False)
-    analyzer = Column(String(20), nullable=False)
-
-class StatsCounters(db.Model):
-    __tablename__ = 'stats_counters'
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    hour = Column(DateTime(timezone=True), nullable=False)
-    packets_count = Column(BigInteger, nullable=False, default=0)
-    joins_count = Column(BigInteger, nullable=False, default=0)
-    alerts_count = Column(BigInteger, nullable=True)
-    devices_count = Column(BigInteger, nullable=True)
-    organization_id = Column(BigInteger, ForeignKey("organization.id"), nullable=False)
-    data_collector_id = Column(BigInteger, ForeignKey("data_collector.id"))
-
-    @classmethod
-    def find(cls, organization_id, since, until, data_collectors):
-        query = cls.group_by_hour(cls, organization_id, since, until, data_collectors)
-        return query.all()
-
-    def group_by_hour(cls, organization_id, since, until, data_collectors):
-        query = cls.query.filter(cls.organization_id == organization_id)
-        if since:
-            query = query.filter(cls.hour >= since)
-        if until:
-            query = query.filter(cls.hour <= until)
-        if data_collectors and len(data_collectors) > 0:
-            query = query.filter(cls.data_collector_id.in_(data_collectors))
-        query = query.with_entities(cast(func.sum(cls.alerts_count), Integer).label('alerts_count'),
-                                    cast(func.sum(cls.packets_count), Integer).label('packets_count'),
-                                    cast(func.sum(cls.devices_count), Integer).label('devices_count'),
-                                    cast(func.sum(cls.joins_count), Integer).label('joins_count'),
-                                    cls.hour)
-        query = query.group_by(cls.hour)
-        query = query.order_by(desc(cls.hour))
-        return query
-
-    @classmethod
-    def group_by_date(cls, organization_id, since, until, data_collectors):
-        query = db.session.query(func.date(cls.hour).label('date'),
-                                 cast(func.sum(cls.alerts_count), Integer).label('alerts_count'),
-                                 cast(func.sum(cls.packets_count), Integer).label('packets_count'),
-                                 cast(func.sum(cls.joins_count), Integer).label('joins_count'))\
-            .filter(cls.organization_id == organization_id)
-
-        if since:
-            query = query.filter(cls.hour >= since)
-        if until:
-            query = query.filter(cls.hour <= until)
-        if data_collectors and len(data_collectors) > 0:
-            query = query.filter(cls.data_collector_id.in_(data_collectors))
-        query = query.group_by(func.date(cls.hour)).order_by(desc(func.date(cls.hour)))
-        return query.all()
-
-    @classmethod
-    def max_devices_by_date(cls, organization_id, since, until, data_collectors):
-        sum_query = cls.group_by_hour(cls, organization_id, since, until, data_collectors).subquery()
-        max_query = sum_query.select().with_only_columns([func.date(sum_query.c.hour).label('date'), func.max(sum_query.c.devices_count).label('max_devices')])
-        query = max_query.group_by(max_query.c.date).order_by(max_query.c.date)
-        return db.session.execute(query).fetchall()
-
-    def to_json_for_packet(self):
-        return {
-            'hour': self.hour,
-            'count': self.packets_count
-        }
-
-    def to_json_for_join(self):
-        return {
-            'hour': self.hour,
-            'count': self.joins_count
-        }
-
-    def to_json_for_device(self):
-        return {
-            'hour': self.hour,
-            'count': self.devices_count
-        }
-
-
 class Params(db.Model):
     __tablename__ = 'params'
     id = Column(Integer, primary_key=True, autoincrement=True)
@@ -1251,63 +1026,9 @@ class Params(db.Model):
         return cls.query.one().url_base
 
 
-class SendMailAttempts(db.Model):
-    __tablename__ = "send_email_attempts"
-
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("iot_user.id"), nullable=False)
-    attempts = Column(Integer, nullable=False, default=True)
-    # last_attempt = db.Column(db.DateTime(timezone=True), nullable=False)
-
-    def to_json(self):
-        return {
-            'id': self.id,
-            'user_id': self.user_id,
-            'attempts': self.attempts
-            # 'last_attempt': self.last_attempt
-        }
-
-    def save_to_db(self):
-        db.session.add(self)
-        db.session.commit()
-
-    def update_to_db(self):
-        db.session.commit()
-
-    def delete_from_db(self):
-        db.session.delete(self)
-        db.session.commit()
-
-    @classmethod
-    def find_by_user(cls, user_id):
-        return cls.query.filter_by(user_id=user_id).first()
-
-
 class QuarantineResolutionReasonType(Enum):
     MANUAL = 'MANUAL'
     AUTOMATIC = 'AUTOMATIC'
-
-
-class GlobalData(db.Model):
-    __tablename__ = "global_data"
-
-    key = Column(String(120),nullable=False, primary_key=True)
-    value = Column(String(300), nullable=False)
-
-    def save_to_db(self):
-        db.session.add(self)
-        db.session.commit()
-
-    def update_to_db(self):
-        db.session.commit()
-
-    def delete_from_db(self):
-        db.session.delete(self)
-        db.session.commit()
-
-    @classmethod
-    def find_by_key(cls, key):
-        return cls.query.filter_by(key=key).first()
 
 
 class QuarantineResolutionReason(db.Model):
