+++ iot_api/user_api/models/__init__.py
@@ -18,3 +18,10 @@ from iot_api.user_api.models.DeviceToTag import DeviceToTag
 from iot_api.user_api.models.GatewayToTag import GatewayToTag
 from iot_api.user_api.models.NotificationAssetTag import NotificationAssetTag
 from iot_api.user_api.models.AppKey import AppKey
+from iot_api.user_api.models.ChangeEmailRequests import ChangeEmailRequests
+from iot_api.user_api.models.GlobalData import GlobalData
+from iot_api.user_api.models.LoginAttempts import LoginAttempts
+from iot_api.user_api.models.PasswordReset import PasswordReset
+from iot_api.user_api.models.RevokedTokenModel import RevokedTokenModel
+from iot_api.user_api.models.SendMailAttempts import SendMailAttempts
+from iot_api.user_api.models.StatsCounters import StatsCounters
